% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combat_ref.R
\name{ComBat_ref}
\alias{ComBat_ref}
\title{Correct batch effects for RNA-seq count data using a negative binomial distribution
Adjust count towards the reference batch with minimum estimated dispersion}
\usage{
ComBat_ref(
  counts,
  batch,
  group = NULL,
  full_mod = TRUE,
  gene.subset.n = NULL,
  genewise.disp = FALSE
)
}
\arguments{
\item{counts}{Raw count matrix from genomic studies (dimensions gene x sample)}

\item{batch}{Batch covariate (only one batch allowed)}

\item{group}{Vector / factor for condition of interest}

\item{full_mod}{Boolean, if TRUE include condition of interest in model}

\item{gene.subset.n}{Number of genes to use in empirical Bayes estimation, only useful when shrink = TRUE}

\item{genewise.disp}{Compute dispersion parameter for each gene = FALSE}
}
\value{
data A probe x sample count matrix, adjusted for batch effects.
}
\description{
Correct batch effects for RNA-seq count data using a negative binomial distribution
Adjust count towards the reference batch with minimum estimated dispersion
}
\examples{
#combatseq_new <- ComBat_ref(counts=cts_sub, batch=batch_sub, group=group_sub, genewise.disp=FALSE)
#adj_counts <- ComBat_ref(counts=cts, batch=batch, group=group,  genewise.disp=FALSE)

}
